backend/pyproject.toml
[tool.poetry]
name = "jktech-assignment"
version = "0.1.0"
description = ""
authors = ["Abhishek Kumar Seth"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = { version = "^0.20.0", extras = ["standard"] }
pandas = "^2.2.2"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.32"}
loguru = "^0.6.0"
python-dotenv = "^1.0.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
yarl = "^1.9.4"
pytest = "^8.3.2"
toml = "^0.10.2"
taskiq = "^0.11.6"
taskiq-fastapi = "^0.3.1"
opentelemetry-api = "^1.26.0"
ujson = "^5.10.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "4.0.1"
ollama = "^0.3.1"
pydantic = "^2.8.2"
langchain = "^0.2.14"
langchain-ollama = "^0.1.1"
pypdf2 = "^3.0.1"
pdfminer = "^20191125"
langchain-core = "^0.2.33"
transformers = "^4.44.0"
langchain-community = "^0.2.12"
pydantic-settings = "^2.4.0"
azure-storage-blob = "^12.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
